builder-agent: separate BuildKit root setup
